fix: futher poi and mission refactoring
