+++ backend/starlink-location/app/api/pois/stats.py
@@ -128,7 +128,7 @@ async def get_next_destination(
         status_eta_mode = "estimated"
         status_phase = None
         try:
-            from app.services.flight_state_manager import get_flight_state_manager
+            from app.services.flight_state import get_flight_state_manager
 
             snapshot = get_flight_state_manager().get_status()
             status_eta_mode = snapshot.eta_mode.value
@@ -159,7 +159,7 @@ async def get_next_destination(
     status_eta_mode = "estimated"
     status_phase = None
     try:
-        from app.services.flight_state_manager import get_flight_state_manager
+        from app.services.flight_state import get_flight_state_manager
 
         snapshot = get_flight_state_manager().get_status()
         status_eta_mode = snapshot.eta_mode.value
@@ -235,7 +235,7 @@ async def get_next_eta(
         status_eta_mode = "estimated"
         status_phase = None
         try:
-            from app.services.flight_state_manager import get_flight_state_manager
+            from app.services.flight_state import get_flight_state_manager
 
             snapshot = get_flight_state_manager().get_status()
             status_eta_mode = snapshot.eta_mode.value
@@ -263,7 +263,7 @@ async def get_next_eta(
     status_eta_mode = "estimated"
     status_phase = None
     try:
-        from app.services.flight_state_manager import get_flight_state_manager
+        from app.services.flight_state import get_flight_state_manager
 
         snapshot = get_flight_state_manager().get_status()
         status_eta_mode = snapshot.eta_mode.value
