+++ backend/starlink-location/tests/unit/test_flight_state_manager.py
@@ -8,7 +8,7 @@ import pytest
 
 from app.models.flight_status import ETAMode, FlightPhase
 from app.models.route import ParsedRoute, RouteMetadata, RouteTimingProfile, RoutePoint
-from app.services.flight_state_manager import get_flight_state_manager
+from app.services.flight_state import get_flight_state_manager
 
 
 @pytest.fixture
